Cleaned up the code and provided comments for all functions and modules.
Just realized, this needs to support fingers with configurable mark and space
in order to support double-thick walls. Need to rework the finger code for
this and it might not resemble the current code at all, so I'm committing this
as it is in case I break everything with the mark-space changes.